+++ R/half.life.R
@@ -1,5 +1,6 @@
 #' Compute the half-life and associated parameters
 #' 
+#' @details 
 #' If \code{manually.selected.points} is \code{FALSE} (default), the
 #' half-life is calculated by computing the best fit line for all 
 #' available sets of points.  The best one is chosen by the following 
@@ -69,19 +70,20 @@
 pk.calc.half.life <- function(conc, time, tmax, tlast,
                               manually.selected.points=FALSE,
                               options=list(),
-                              min.hl.points=PKNCA.choose.option("min.hl.points", options),
-                              adj.r.squared.factor=PKNCA.choose.option("adj.r.squared.factor", options),
-                              conc.blq=PKNCA.choose.option("conc.blq", options),
-                              conc.na=PKNCA.choose.option("conc.na", options),
-                              first.tmax=PKNCA.choose.option("first.tmax", options),
-                              allow.tmax.in.half.life=PKNCA.choose.option("allow.tmax.in.half.life", options),
+                              min.hl.points=NULL,
+                              adj.r.squared.factor=NULL,
+                              conc.blq=NULL,
+                              conc.na=NULL,
+                              first.tmax=NULL,
+                              allow.tmax.in.half.life=NULL,
                               check=TRUE) {
   ## Check inputs
-  min.hl.points <-
-    PKNCA.options(min.hl.points=min.hl.points, check=TRUE)
-  adj.r.squared.factor <-
-    PKNCA.options(adj.r.squared.factor=adj.r.squared.factor,
-                  check=TRUE)
+  min.hl.points <- PKNCA.choose.option(name="min.hl.points", value=min.hl.points, options=options)
+  adj.r.squared.factor <- PKNCA.choose.option(name="adj.r.squared.factor", value=adj.r.squared.factor, options=options)
+  conc.blq <- PKNCA.choose.option(name="conc.blq", value=conc.blq, options=options)
+  conc.na <- PKNCA.choose.option(name="conc.na", value=conc.na, options=options)
+  first.tmax <- PKNCA.choose.option(name="first.tmax", value=first.tmax, options=options)
+  allow.tmax.in.half.life <- PKNCA.choose.option(name="allow.tmax.in.half.life", value=allow.tmax.in.half.life, options=options)
   if (check) {
     check.conc.time(conc, time)
     data <- clean.conc.blq(conc, time, conc.blq=conc.blq, conc.na=conc.na)
@@ -229,11 +231,13 @@ add.interval.col("r.squared",
                  values=c(FALSE, TRUE),
                  desc="The r^2 value of the half-life calculation",
                  depends=c("half.life"))
+PKNCA.set.summary("r.squared", business.mean, business.sd)
 add.interval.col("adj.r.squared",
                  FUN=NA,
                  values=c(FALSE, TRUE),
                  desc="The adjusted r^2 value of the half-life calculation",
                  depends=c("half.life"))
+PKNCA.set.summary("adj.r.squared", business.mean, business.sd)
 add.interval.col("lambda.z",
                  FUN=NA,
                  values=c(FALSE, TRUE),
@@ -245,6 +249,7 @@ add.interval.col("lambda.z.time.first",
                  values=c(FALSE, TRUE),
                  desc="The first time point used for the calculation of half-life",
                  depends=c("half.life"))
+PKNCA.set.summary("lambda.z.time.first", business.median, business.range)
 add.interval.col("lambda.z.n.points",
                  FUN=NA,
                  values=c(FALSE, TRUE),