+++ R/pk.calc.simple.R
@@ -81,8 +81,9 @@ PKNCA.set.summary("cmin", business.geomean, business.geocv)
 #' @export
 pk.calc.tmax <- function(conc, time,
                          options=list(),
-                         first.tmax=PKNCA.choose.option("first.tmax", options),
+                         first.tmax=NULL,
                          check=TRUE) {
+  first.tmax <- PKNCA.choose.option(name="first.tmax", value=first.tmax, options=options)
   if (missing(conc))
     stop("conc must be given")
   if (missing(time))
@@ -538,7 +539,7 @@ add.interval.col("mrt.md.pred",
                  desc="The mean residence time with multiple dosing and nonlinear kinetics using predicted Clast",
                  formalsmap=list(auctau="auclast", aumctau="aumclast", aucinf="aucinf.pred"),
                  depends=c("auclast", "aumclast", "aucinf.pred"))
-PKNCA.set.summary("mrt.md.obs", business.geomean, business.geocv)
+PKNCA.set.summary("mrt.md.pred", business.geomean, business.geocv)
 
 #' Calculate the terminal volume of distribution (Vz)
 #'