+++ tests/testthat/test-interpolate.conc.R
@@ -322,7 +322,11 @@ test_that("interpolate.conc", {
                                 time=0:1,
                                 time.out=0.5,
                                 interp.method="this doesn't work"),
-               regexp="interp.method must be one of 'linear' or 'lin up/log down'",
+               regexp=tryCatch(expr={
+                 match.arg("foo", choices=c("lin up/log down", "linear"))
+               },
+               error=function(e) e)$message,
+               fixed=TRUE,
                info="Confirm that invalid interpolation methods are an error.")
 
   expect_error(interpolate.conc(conc=0:1,
@@ -585,4 +589,20 @@ test_that("interp.extrap.conc.dose", {
                                        duration.dose=0),
                structure(2, Method="After an IV bolus with a concentration next"),
                info="After IV bolus, one concentration")
+  
+  expect_equal(interp.extrap.conc.dose(conc=c(0, 1, 2, 1, 0.5, 0.25),
+                                       time=c(-1, 1:5),
+                                       time.dose=0,
+                                       time.out=c(-2, 2)),
+               structure(c(0, 2),
+                         Method=c("Before all events", "Observed concentration")),
+               info="Outputs are in the same order as inputs (initially sorted)")
+  
+  expect_equal(interp.extrap.conc.dose(conc=c(0, 1, 2, 1, 0.5, 0.25),
+                                       time=c(-1, 1:5),
+                                       time.dose=0,
+                                       time.out=c(2, -2)),
+               structure(c(2, 0),
+                         Method=c("Observed concentration", "Before all events")),
+               info="Outputs are in the same order as inputs (reverse sorted time.out)")
 })